+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/TestGuiceyApp.java
@@ -204,10 +204,24 @@ public @interface TestGuiceyApp {
      * may be required in case when custom client configuration is required for test.
      * <p>
      * Core test does not start dropwizard web services, but client still could be used to call external services.
+     * <p>
+     * Note: value is ignored when {@link #useApacheClient()} set to true
      *
      * @return client factory class
      */
     Class<? extends TestClientFactory> clientFactory() default DefaultTestClientFactory.class;
+
+    /**
+     * Shortcut for {@link #clientFactory()} to configure
+     * {@link ru.vyarus.dropwizard.guice.test.client.ApacheTestClientFactory}. The default
+     * {@link org.glassfish.jersey.client.HttpUrlConnectorProvider} supports only HTTP 1.1 methods and have
+     * problem with PATCH method usage on jdk > 16.
+     * <p>
+     * Note: {@link #clientFactory()} value is ignored when set to true,
+     *
+     * @return true to use apache connection provider in jersey client
+     */
+    boolean useApacheClient() default false;
 
     /**
      * By default, guicey simulates {@link io.dropwizard.lifecycle.Managed} objects lifecycle.